Trimmed library
